fix re-export types in translation
