+++ packages/superset-ui-connection/src/index.ts
@@ -1,4 +1,4 @@
 export { default as callApi } from './callApi';
-export { default as SupersetClient, SupersetClientInterface } from './SupersetClient';
-export { SupersetClientClass } from './SupersetClientClass';
+export { default as SupersetClient } from './SupersetClient';
+export { default as SupersetClientClass } from './SupersetClientClass';
 export * from './types';
