+++ packages/superset-ui-core/src/index.ts
@@ -1,13 +1,12 @@
 export { default as ExtensibleFunction } from './models/ExtensibleFunction';
 export { default as Plugin } from './models/Plugin';
-export { default as Preset, PresetConfig } from './models/Preset';
-export { default as Registry, RegistryConfig, OverwritePolicy } from './models/Registry';
-export {
-  default as RegistryWithDefaultKey,
-  RegistryWithDefaultKeyConfig,
-} from './models/RegistryWithDefaultKey';
+export { default as Preset } from './models/Preset';
+export { default as Registry } from './models/Registry';
+export { default as RegistryWithDefaultKey } from './models/RegistryWithDefaultKey';
 
 export { default as convertKeysToCamelCase } from './utils/convertKeysToCamelCase';
 export { default as isDefined } from './utils/isDefined';
 export { default as isRequired } from './utils/isRequired';
 export { default as makeSingleton } from './utils/makeSingleton';
+
+export * from './types';
