+++ packages/superset-ui-core/src/models/Preset.ts
@@ -1,19 +1,19 @@
 import Plugin from './Plugin';
 
-export interface PresetConfig {
-  name?: string;
-  description?: string;
-  presets?: Preset[];
-  plugins?: Plugin[];
-}
-
 export default class Preset {
   name: string;
   description: string;
   presets: Preset[];
   plugins: Plugin[];
 
-  constructor(config: PresetConfig = {}) {
+  constructor(
+    config: {
+      name?: string;
+      description?: string;
+      presets?: Preset[];
+      plugins?: Plugin[];
+    } = {},
+  ) {
     const { name = '', description = '', presets = [], plugins = [] } = config;
     this.name = name;
     this.description = description;
