+++ packages/superset-ui-core/src/types.ts
@@ -1,0 +1,5 @@
+export enum OverwritePolicy {
+  ALLOW = 'ALLOW',
+  PROHIBIT = 'PROHIBIT',
+  WARN = 'WARN',
+}
