+++ packages/superset-ui-translation/src/index.ts
@@ -1,2 +1,2 @@
 export { configure, t, tn } from './TranslatorSingleton';
-export { TranslatorConfig, LanguagePack } from './Translator';
+export * from './types';
