+++ packages/superset-ui-translation/src/types.ts
@@ -20,3 +20,7 @@ export interface Jed {
   ifPlural(value: number, plural: string): Jed;
   fetch(...args: any[]): string;
 }
+
+export interface TranslatorConfig {
+  languagePack?: LanguagePack;
+}
